Moved request_succeeded and its handled from engine into rq_handler
- Migrated clear rq_handler logic from engine into redis_handler
- Updated sender to be the class that produces the job
