+++ cvat/apps/redis_handler/utils.py
@@ -3,7 +3,9 @@
 # SPDX-License-Identifier: MIT
 
 import importlib
+import traceback
 from pathlib import Path
+from typing import Any, Self
 
 import rq
 from rq.job import Job as RQJob
@@ -15,8 +17,63 @@ def get_class_from_module(module_path: str | Path, class_name: str) -> type | No
     return klass
 
 
+def get_class_from_full_path(full_path: str) -> type | None:
+    module_path, class_name = full_path.rsplit(".", 1)
+    return get_class_from_module(module_path, class_name)
+
+
 def rq_job_will_be_retried(rq_job: RQJob) -> bool:
     return bool(rq_job.retries_left and rq_job.retries_left > 0)
+
+
+def send_request_succeeded_signal(
+    rq_job: RQJob,
+    connection: Any,
+    result: Any,
+) -> None:
+    from cvat.apps.engine.rq import BaseRQMeta
+
+    request_manager_cls_path = BaseRQMeta.for_job(rq_job).request_manager_cls
+    sender = (
+        get_class_from_full_path(full_path=request_manager_cls_path)
+        if request_manager_cls_path
+        else None
+    )
+
+    _ = signals.request_succeeded.send_robust(
+        sender=sender,
+        rq_job=rq_job,
+        status=JobStatus.FINISHED,
+        message=None,
+    )
+
+
+def send_request_failed_signal(
+    rq_job: RQJob,
+    connection: Any,
+    exc_type: type[BaseException],
+    exc_value: BaseException,
+    exc_traceback: Any,
+) -> None:
+    from cvat.apps.engine.rq import BaseRQMeta
+    from cvat.apps.engine.utils import parse_exception_message
+
+    if rq_job_will_be_retried(rq_job=rq_job):
+        return
+
+    request_manager_cls_path = BaseRQMeta.for_job(rq_job).request_manager_cls
+    sender = (
+        get_class_from_full_path(request_manager_cls_path) if request_manager_cls_path else None
+    )
+
+    _ = signals.request_failed.send_robust(
+        sender=sender,
+        rq_job=rq_job,
+        status=JobStatus.FAILED,
+        message=parse_exception_message(
+            "".join(traceback.format_exception_only(exc_type, exc_value))
+        ),
+    )
 
 
 def get_current_job_attempt() -> int:
